requirements.txt
Flask==2.3.2
torch==2.0.1
Pillow==10.0.0
cn_clip==2.0.0
moviepy==1.0.3
numpy==1.24.4
requests==2.31.0
sqlite3==0.1.0
